% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworks.R
\name{getFollowerIDs}
\alias{getFollowerIDs}
\title{Downloading the follower ids of a list of users}
\usage{
getFollowerIDs(
  screen_name_list = NULL,
  user_id_list = NULL,
  auth_df,
  sleepTime = 30,
  max_per_user = 15000,
  verbose = F
)
}
\arguments{
\item{screen_name_list}{A list of Twitter screen names}

\item{user_id_list}{A list of Twitter numerical ids (as character)}

\item{auth_df}{The dataframe containing the Twitter keys (check read.keys function)}

\item{sleepTime}{Sleep time (in seconds) when activating next key}

\item{max_per_user}{Max number of followers to download per user}

\item{verbose}{Monitor the process (recommended for debuging!)}
}
\value{
A list of users and their followers
}
\description{
This function takes a single or a list of users (user id or user screen name) and downloads the followers of each user.
Either of screen_name_list or user_id_list must be provided
}
