% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworks.R
\name{getFriendFollowersIDs}
\alias{getFriendFollowersIDs}
\title{Downloading the friends' and folowers' ids of a list of users}
\usage{
getFriendFollowersIDs(
  screen_name_list = NULL,
  user_id_list = NULL,
  auth_df,
  sleepTime = 30,
  max_per_user = 15000,
  verbose = F,
  jsonPath = NULL,
  namePatternFriends = "fri-",
  namePatternFollowers = "fol-"
)
}
\arguments{
\item{screen_name_list}{A list of Twitter screen names}

\item{user_id_list}{A list of Twitter numerical ids (as character)}

\item{auth_df}{The dataframe containing the Twitter keys (check read.keys function)}

\item{sleepTime}{Sleep time (in seconds) when activating next key}

\item{max_per_user}{Max number of friends to download per user}

\item{verbose}{Monitor the process (recommended for debuging!)}

\item{jsonPath}{Where to save the json files}
}
\value{
A list of users and their friends
}
\description{
This function takes a single or a list of users (user id or user screen name) and downloads the friends of each user.
Either of screen_name_list or user_id_list must be provided
}
