% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{read.keys}
\alias{read.keys}
\title{Loading the csv file containing the list of Twitter keys}
\usage{
read.keys(file)
}
\arguments{
\item{file}{Path to the file}
}
\value{
A dataframe of Twitter keys
}
\description{
This function takes a file path as input and loads the keys from the file.
Each line (row) should represent a unique key. There csv file should contain no header and the order of the columns should be:
"oauth_consumer_key", "consumer_secret", "oauth_token", "oauth_token_secret"
Files with less or more than four columns are not valid.
}
