% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{showUsers}
\alias{showUsers}
\title{Downloading a single user or alist of users}
\usage{
showUsers(
  screen_name_list = NULL,
  user_id_list = NULL,
  include_entities = T,
  auth_df,
  sleepTime = 30,
  verbose = F
)
}
\arguments{
\item{screen_name_list}{A list of Twitter screen names}

\item{user_id_list}{A list of Twitter numerical ids (as character)}

\item{auth_df}{The dataframe containing the Twitter keys (check read.keys function)}

\item{sleepTime}{Sleep time (in seconds) when activating next key}

\item{verbose}{Monitor the process (recommended for debuging!)}
}
\value{
A data.table of last Tweets of each user (all the columns are returned!)
}
\description{
This function takes a single or a list of users (user id or user screen name) and downloads the last Tweet and informations on the user.
Either of screen_name_list or user_id_list must be provided
}
