% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphAnalysis.R
\name{userEdgeList2graph}
\alias{userEdgeList2graph}
\title{Converting a list of users containing their friends or foloowrs to an igraph network}
\usage{
userEdgeList2graph(..., directed)
}
\arguments{
\item{directed}{to create a directed network or indirected}
}
\value{
An igraph network
}
\description{
This function takes multiple lists as input. Each element of the list is a user
and each element contains the list of friend or follower ids.
}
